Added randomization machinery
Both the question order and the order of the answers for each question can now be randomized independently. The order of the answers is re-randomized each time they are asked, which seems most challenging.
This was a very large amount of work involving major changes to the way the web application is structured internally. Also, the randomization state had to be stored in the session to make the answer evaluation page work correctly.