Fixed bug due to hard-coded base URL
The problem is that the base URL (like /grolprep/bin/fcc-grol-prep.cgi), that would be prepended to URLs for all sorts of things like form actions, was hard-coded to be /grolprep/
So suppose the app is installed to another dir (like grolprep-version). After the first request-response it starts referencing a totally different installation dir (if it's even there). A terrible mixing of the two installations ensues, very bad!
Now the code is computing the URL prefix using the generated Paths_grolprep module.