Fixed bad session due to newlines in question list bug
This was a tricky problem to find. Finally Bill (a customer) realized that he was seeing this damaged session when there was a newline delimiting question numbers in the custom question input field.

Oddly, we were splitting on spaces only (wordsBy (== ' ')), not just on all whitespace (as the words function will do).

Fixed