Reorganized source files into app and lib dirs
This was done because the `src/data-import.hs` script was becoming
difficult to run with its dependencies not installed somewhere.

Also modified the db-regen.sh script which builds the database using
data-import.

As part of this change, the project is now using hpack to manage the
cabal config.

Some minor warnings were also addressed.